moved logic from task controller to the task service

+++ src/main/java/com/alexrmn/todolistspringboot/service/TaskService.java
@@ -2,12 +2,17 @@ package com.alexrmn.todolistspringboot.service;
 
 import com.alexrmn.todolistspringboot.exception.BusinessException;
 import com.alexrmn.todolistspringboot.model.Task;
+import com.alexrmn.todolistspringboot.model.User;
 import com.alexrmn.todolistspringboot.model.dto.CreateTaskDto;
 import com.alexrmn.todolistspringboot.model.dto.UpdateTaskDto;
 import com.alexrmn.todolistspringboot.repository.TaskRepository;
+import com.alexrmn.todolistspringboot.repository.UserRepository;
+import com.alexrmn.todolistspringboot.util.AuthUtils;
 import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -17,14 +22,21 @@ import java.util.List;
 public class TaskService {
 
     private final TaskRepository taskRepository;
+    private final UserRepository userRepository;
+
 
     public List<Task> findAll(){
         return taskRepository.findAll();
     }
 
     public Task findById(Integer id){
-        return taskRepository.findById(id)
-                .orElseThrow(() -> new BusinessException(HttpStatus.NO_CONTENT, "Task with id " + id + " wasn't found."));
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        Task task = taskRepository.findById(id)
+                .orElseThrow(() -> new BusinessException(HttpStatus.BAD_REQUEST, "Task doesn't exist"));
+        if (!AuthUtils.taskBelongsToUser(task, authentication) && !AuthUtils.isAdmin(authentication)) {
+            throw new BusinessException(HttpStatus.FORBIDDEN, "You are not authorised to see this task");
+        }
+        return task;
 
     }
 
@@ -33,15 +45,32 @@ public class TaskService {
     }
 
     public List<Task> findByUserId(Integer id){
+        User user = userRepository.findById(id)
+                .orElseThrow(() -> new BusinessException(HttpStatus.BAD_REQUEST, "User doesn't exist"));
+
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if (!user.getId().equals(AuthUtils.getLoggedInUser(authentication).getId())) {
+            throw new BusinessException(HttpStatus.FORBIDDEN, "User not authorized");
+        }
+
         return taskRepository.findByUserId(id);
     }
 
     public void saveTask(CreateTaskDto createTaskDto){
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        createTaskDto.setUser(AuthUtils.getLoggedInUser(authentication));
         taskRepository.save(new Task(createTaskDto));
     }
 
     public void deleteTask(Task task) {
-        taskRepository.findById(task.getId()).orElseThrow(() -> new BusinessException(HttpStatus.NO_CONTENT, "Task not found"));
+        taskRepository.findById(task.getId())
+                .orElseThrow(() -> new BusinessException(HttpStatus.NO_CONTENT, "Task not found"));
+
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if (!task.getUser().getId().equals(AuthUtils.getLoggedInUser(authentication).getId())) {
+            throw new BusinessException(HttpStatus.FORBIDDEN, "User not authorized");
+        }
+
         taskRepository.delete(task);
     }
